Reorganize universal I/O code files
