+++ lib/common/common/src/universal_io/disk_cache/mod.rs
@@ -4,19 +4,21 @@ use std::path::{Path, PathBuf};
 
 use fs_err as fs;
 
-use crate::generic_consts::{AccessPattern, Random};
+use crate::generic_consts::AccessPattern;
 use crate::universal_io::{
     OpenOptions, ReadRange, Result, UniversalIoError, UniversalRead, UniversalReadFileOps,
-    UniversalReadPipeline, UserData, local_file_ops,
+    UserData, local_file_ops,
 };
 
 mod cached_slice;
 mod controller;
+mod pipeline;
 #[cfg(test)]
 mod tests;
 
 pub use cached_slice::CachedSlice;
 use controller::{CacheController, CacheRead};
+use pipeline::DiskCacheReadPipeline;
 
 use super::UniversalKind;
 
@@ -120,50 +122,5 @@ impl UniversalRead for CachedSlice {
 
     fn kind() -> UniversalKind {
         UniversalKind::DiskCache
-    }
-}
-
-pub struct DiskCacheReadPipeline<'file, T, U>
-where
-    T: bytemuck::Pod,
-    U: UserData,
-{
-    result: Option<(U, Cow<'file, [T]>)>,
-}
-
-impl<'file, T, U> UniversalReadPipeline<'file, T, U> for DiskCacheReadPipeline<'file, T, U>
-where
-    T: bytemuck::Pod,
-    U: UserData,
-{
-    type File = CachedSlice;
-
-    fn new() -> Result<Self> {
-        Ok(Self { result: None })
-    }
-
-    fn can_schedule(&mut self) -> bool {
-        self.result.is_none()
-    }
-
-    fn schedule<P>(
-        &mut self,
-        user_data: U,
-        file: &'file CachedSlice,
-        range: ReadRange,
-    ) -> Result<()>
-    where
-        P: AccessPattern,
-    {
-        if self.result.is_some() {
-            return Err(UniversalIoError::QueueIsFull);
-        }
-
-        self.result = Some((user_data, file.read::<Random, T>(range)?));
-        Ok(())
-    }
-
-    fn wait(&mut self) -> Result<Option<(U, Cow<'file, [T]>)>> {
-        Ok(self.result.take())
     }
 }
