+++ lib/common/common/src/universal_io/io_uring/mod.rs
@@ -1,4 +1,5 @@
 mod error;
+mod pipeline;
 mod pool;
 mod runtime;
 
@@ -8,7 +9,7 @@ mod tests;
 use std::borrow::Cow;
 use std::io::{self, Read as _, Seek as _};
 use std::os::fd::AsRawFd as _;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
 use ::io_uring::types::Fd;
@@ -16,8 +17,10 @@ use fs_err as fs;
 use fs_err::os::unix::fs::{FileExt as _, OpenOptionsExt as _};
 
 use self::error::*;
+use self::pipeline::IoUringPipeline;
 use self::pool::*;
 use self::runtime::*;
+use super::traits::UniversalReadFileOps;
 use super::*;
 use crate::generic_consts::AccessPattern;
 
@@ -29,11 +32,11 @@ pub struct IoUringFile {
     ///
     /// This is because `O_DIRECT` can only read in aligned blocks of data, so reads at EOF might not
     /// be aligned with O_DIRECT alignment, but it is not possible to request less than one block.
-    direct_io: bool,
+    pub(super) direct_io: bool,
 }
 
 impl IoUringFile {
-    fn fd(&self) -> Fd {
+    pub(super) fn fd(&self) -> Fd {
         Fd(self.file.as_raw_fd())
     }
 }
@@ -139,79 +142,6 @@ impl UniversalRead for IoUringFile {
 
     fn kind() -> UniversalKind {
         UniversalKind::IoUring
-    }
-}
-
-pub struct IoUringPipeline<'file, T, U>
-where
-    T: bytemuck::Pod,
-    U: UserData,
-{
-    runtime: IoUringRuntime<'file, T, U>,
-}
-
-impl<'file, T, U> UniversalReadPipeline<'file, T, U> for IoUringPipeline<'file, T, U>
-where
-    T: bytemuck::Pod,
-    U: UserData,
-{
-    type File = IoUringFile;
-
-    fn new() -> Result<Self> {
-        Ok(Self {
-            runtime: IoUringRuntime::new()?,
-        })
-    }
-
-    fn can_schedule(&mut self) -> bool {
-        let squeue = self.runtime.io_uring.submission();
-        self.runtime.in_progress + squeue.len() < IO_URING_QUEUE_LENGTH as _
-    }
-
-    fn schedule<P>(
-        &mut self,
-        user_data: U,
-        file: &'file IoUringFile,
-        range: ReadRange,
-    ) -> Result<()>
-    where
-        P: AccessPattern,
-    {
-        let mut squeue = self.runtime.io_uring.submission();
-
-        if self.runtime.in_progress + squeue.len() >= IO_URING_QUEUE_LENGTH as _ {
-            return Err(UniversalIoError::QueueIsFull);
-        }
-
-        let entry = self
-            .runtime
-            .state
-            .read(user_data, file.fd(), range, file.direct_io);
-
-        unsafe {
-            squeue.push(&entry).expect("submission queue is not full");
-        }
-
-        Ok(())
-    }
-
-    fn wait(&mut self) -> Result<Option<(U, Cow<'file, [T]>)>> {
-        let next = self.runtime.completed().next();
-
-        let enqueued = self.runtime.enqueued();
-
-        if next.is_some() && enqueued > 0 {
-            self.runtime.submit_and_wait(0)?;
-        } else if next.is_none() && enqueued + self.runtime.in_progress > 0 {
-            self.runtime.submit_and_wait(1)?;
-        }
-
-        let Some(result) = next.or_else(|| self.runtime.completed().next()) else {
-            return Ok(None);
-        };
-
-        let (user_data, resp) = result?;
-        Ok(Some((user_data, Cow::Owned(resp.expect_read()))))
     }
 }
