+++ lib/common/common/src/universal_io/traits/read.rs
@@ -1,8 +1,9 @@
 use std::borrow::Cow;
 use std::path::Path;
 
-use super::*;
+use super::{UniversalReadFileOps, UniversalReadPipeline, UserData};
 use crate::generic_consts::{AccessPattern, Sequential};
+use crate::universal_io::{OpenOptions, ReadRange, Result, UniversalKind};
 
 /// Interface for accessing files in a universal way, abstracting away possible
 /// implementations, such as memory map, io_uring, DIRECTIO, S3, etc.
@@ -134,48 +135,3 @@ pub trait UniversalRead: UniversalReadFileOps {
 
     // When adding provided methods, don't forget to update impls in crate::universal_io::wrappers::*.
 }
-
-pub trait UniversalReadPipeline<'file, T, U>: Sized
-where
-    T: bytemuck::Pod,
-    U: UserData,
-{
-    type File: 'file;
-
-    fn new() -> Result<Self>;
-
-    fn can_schedule(&mut self) -> bool;
-
-    /// Schedule read operation.
-    ///
-    /// Note: an implementation might add it to internal queue, but not actually
-    /// execute it until [`UniversalReadPipeline::wait()`] is called.
-    ///
-    /// Should be called only when [`UniversalReadPipeline::can_schedule()`] is
-    /// `true`. Returns [`UniversalIoError::QueueIsFull`] otherwise.
-    fn schedule<P>(
-        &mut self,
-        user_data: U,
-        file: &'file Self::File,
-        range: ReadRange,
-    ) -> Result<()>
-    where
-        P: AccessPattern;
-
-    /// Block until any of the scheduled operations is completed and consume its
-    /// result.
-    fn wait(&mut self) -> Result<Option<(U, Cow<'file, [T]>)>>;
-}
-
-/// An arbitrary value to distinguish requests.
-///
-/// Batched universal I/O methods let callers to add an arbitrary user-provided
-/// value to each request. This value will be passed back to the caller/callback
-/// when the request completes.
-///
-/// Similar to `user_data` in `io_uring`, but allows arbitrary type, not just
-/// `u64`.
-///
-/// This trait exists for documentation/code navigation purposes only.
-pub trait UserData {}
-impl<T> UserData for T {}
